server_mode/client/requirements.txt
python-dotenv
pyperclip
requests
